+++ inc/apiclassic.class.php
@@ -76,6 +76,90 @@
     }
 
     /**
+     * Add a new item through a JSS Classic API endpoint.
+     * @since 1.1.0
+     * @param string  $endpoint The API endpoint.
+     * @param array   $data Associative array of data to post to the endpoint.
+     * @return int|bool True if successful, or the HTTP return code if it is not 201.
+     */
+    private static function add(string $endpoint, array $data)
+    {
+        if (!self::$connection) {
+            self::$connection = new PluginJamfConnection();
+        }
+        $url = (self::$connection)->getAPIUrl($endpoint);
+        $curl = curl_init($url);
+        // Set the username and password in an authentication header
+        self::$connection->setCurlAuth($curl);
+        curl_setopt($curl, CURLOPT_POSTFIELDS, json_encode($data));
+        curl_setopt($curl, CURLOPT_SSLVERSION, 6);
+        curl_setopt($curl, CURLOPT_HTTPHEADER, [
+           'Content-Type: application/json',
+           'Accept: application/json'
+        ]);
+        curl_exec($curl);
+        $httpcode = curl_getinfo($curl, CURLINFO_HTTP_CODE);
+        curl_close($curl);
+        return ($httpcode == 201) ? true : $httpcode;
+    }
+
+    /**
+     * Update an item through a JSS Classic API endpoint.
+     * @since 1.1.0
+     * @param string  $endpoint The API endpoint.
+     * @param array   $data Associative array of data to put to the endpoint.
+     * @return int|bool True if successful, or the HTTP return code if it is not 201.
+     */
+    private static function update(string $endpoint, array $data)
+    {
+        if (!self::$connection) {
+            self::$connection = new PluginJamfConnection();
+        }
+        $url = (self::$connection)->getAPIUrl($endpoint);
+        $curl = curl_init($url);
+        // Set the username and password in an authentication header
+        self::$connection->setCurlAuth($curl);
+        curl_setopt($curl, CURLOPT_CUSTOMREQUEST, 'PUT');
+        curl_setopt($curl, CURLOPT_POSTFIELDS, json_encode($data));
+        curl_setopt($curl, CURLOPT_SSLVERSION, 6);
+        curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, 0);
+        curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, 0);
+        curl_setopt($curl, CURLOPT_HTTPHEADER, [
+           'Content-Type: application/json',
+           'Accept: application/json'
+        ]);
+        curl_exec($curl);
+        $httpcode = curl_getinfo($curl, CURLINFO_HTTP_CODE);
+        curl_close($curl);
+        return ($httpcode == 201) ? true : $httpcode;
+    }
+
+    /**
+     * Delete an item through a JSS Classic API endpoint.
+     * @since 1.1.0
+     * @param string  $endpoint The API endpoint.
+     * @return int|bool True if successful, or the HTTP return code if it is not 200.
+     */
+    private static function delete(string $endpoint)
+    {
+        if (!self::$connection) {
+            self::$connection = new PluginJamfConnection();
+        }
+        $url = (self::$connection)->getAPIUrl($endpoint);
+        $curl = curl_init($url);
+        // Set the username and password in an authentication header
+        self::$connection->setCurlAuth($curl);
+        curl_setopt($curl, CURLOPT_CUSTOMREQUEST, 'DELETE');
+        curl_setopt($curl, CURLOPT_SSLVERSION, 6);
+        curl_setopt($curl, CURLOPT_SSL_VERIFYHOST, 0);
+        curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, 0);
+        curl_exec($curl);
+        $httpcode = curl_getinfo($curl, CURLINFO_HTTP_CODE);
+        curl_close($curl);
+        return ($httpcode == 200) ? true : $httpcode;
+    }
+
+    /**
      * Construct a parameter query string for an API endpoint.
      * @since 1.0.0
      * @param array $params API inputs.
@@ -105,5 +189,52 @@
         // Strip first key (usually like mobile_devices or mobile_device)
         // No other first level keys exist
         return (!is_null($response) && count($response)) ? reset($response) : null;
+    }
+
+    /**
+     * Add an item of the specified JSS itemtype and parameters.
+     * @since 1.1.0
+     * @param string $itemtype The type of data to fetch. This matches up with endpoint names.
+     * @param array $params API input parameters such as udid, name, or subset.
+     * @param array $fields Associative array of item fields.
+     * @return array Associative array of the decoded JSON response.
+     */
+    public static function addItem(string $itemtype, array $params = [], array $fields = [])
+    {
+        $param_str = self::getParamString($params);
+        $endpoint = "$itemtype$param_str";
+        $response = self::add($endpoint, $fields);
+        return $response;
+    }
+
+    /**
+     * Update an item of the specified JSS itemtype and parameters.
+     * @since 1.1.0
+     * @param string $itemtype The type of data to fetch. This matches up with endpoint names.
+     * @param array $params API input parameters such as udid, name, or subset.
+     * @param array $fields Associative array of item fields.
+     * @return array Associative array of the decoded JSON response.
+     */
+    public static function updateItem(string $itemtype, array $params = [], array $fields = [])
+    {
+        $param_str = self::getParamString($params);
+        $endpoint = "$itemtype$param_str";
+        $response = self::update($endpoint, $fields);
+        return $response;
+    }
+
+    /**
+     * Delete an item of the specified JSS itemtype and parameters.
+     * @since 1.1.0
+     * @param string $itemtype The type of data to fetch. This matches up with endpoint names.
+     * @param array $params API input parameters such as udid, name, or subset.
+     * @return array Associative array of the decoded JSON response.
+     */
+    public static function deleteItem(string $itemtype, array $params = [])
+    {
+        $param_str = self::getParamString($params);
+        $endpoint = "$itemtype$param_str";
+        $response = self::delete($endpoint);
+        return $response;
     }
 }