+++ inc/mobiledevice.class.php
@@ -91,10 +91,13 @@ class PluginJamfMobileDevice extends CommonDBChild
       $out .= "<td>".__('Activation locked', 'jamf')."</td>";
       $out .= "<td>".$getYesNo($match['activation_lock_enabled'])."</td></tr>";
 
+      $link = self::getJamfDeviceURL($match['jamf_items_id']);
+      $view_msg = __('View in Jamf', 'jamf');
+      $out .= "<tr><td colspan='4' class='center'>";
+      $out .= "<a class='vsubmit' href='{$link}' target='_blank'>{$view_msg}</a>";
+
       if ($item->canUpdate()) {
-         $out .= "<tr><td colspan='4' class='center'>";
-         $out .= "<a class='vsubmit' onclick='syncDevice(&quot;{$item::getType()}&quot;, {$item->getID()}); return false;'>".__('Sync now', 'jamf')."</a>";
-         $out .= "</td></tr>";
+         $out .= "&nbsp;&nbsp;<a class='vsubmit' onclick='syncDevice(&quot;{$item::getType()}&quot;, {$item->getID()}); return false;'>".__('Sync now', 'jamf')."</a>";
          $ajax_url = $CFG_GLPI['root_doc']."/plugins/jamf/ajax/sync.php";
          $js = <<<JAVASCRIPT
                function syncDevice(itemtype, items_id) {
@@ -111,6 +114,7 @@ class PluginJamfMobileDevice extends CommonDBChild
 JAVASCRIPT;
          $out .= Html::scriptBlock($js);
       }
+      $out .= "</td></tr>";
 
       $out .= "<th colspan='4'>".__('Jamf Lost Mode Information', 'jamf')."</th>";
       $enabled = $match['lost_mode_enabled'];
@@ -198,5 +202,57 @@ JAVASCRIPT;
          'itemtype' => $item::getType(),
          'items_id' => $item->getID()
       ]);
+   }
+
+   /**
+    * @param CommonDBTM $item
+    * @return PluginJamfMobileDevice
+    */
+   public static function getJamfItemForGLPIItem(CommonDBTM $item)
+   {
+       $mobiledevice = new PluginJamfMobileDevice();
+       $matches = $mobiledevice->find([
+           'itemtype'   => $item::getType(),
+           'items_id'   => $item->getID()
+       ], [], 1);
+       if (count($matches)) {
+           $id = reset($matches)['id'];
+           $mobiledevice->getFromDB($id);
+           return $mobiledevice;
+       }
+       return null;
+   }
+
+   public function getExtensionAttributes()
+   {
+       global $DB;
+
+       $ext_table = PluginJamfExtensionAttribute::getTable();
+       $item_ext_table = PluginJamfItem_ExtensionAttribute::getTable();
+
+       $iterator = $DB->request([
+           'SELECT' => [
+               'name', 'data_type', 'value'
+           ],
+           'FROM'   => $ext_table,
+           'LEFT JOIN'  => [
+               $item_ext_table => [
+                   'FKEY'   => [
+                       $ext_table       => 'id',
+                       $item_ext_table  => 'glpi_plugin_jamf_extensionattributes_id'
+                   ]
+               ]
+           ],
+           'WHERE'  => [
+               $item_ext_table.'.itemtype'   => self::getType(),
+               'items_id'   => $this->getID()
+           ]
+       ]);
+
+       $attributes = [];
+       while ($data = $iterator->next()) {
+           $attributes[] = $data;
+       }
+       return $attributes;
    }
 }